setting up separate server/client env modules
