+++ js/env-client.js
@@ -1,0 +1,4 @@
+// Scrape env from dom
+let scrapedEnv = JSON.parse(document.getElementById(`environment-variables`).textContent);
+
+export default scrapedEnv;
